+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## Next release (v0.13.0)
+
+### Enhancements
+
+- Use `uuid` data type for event `correlation_id` and `causation_id` ([#57](https://github.com/slashdotdash/eventstore/pull/57)).
+
+### Bug fixes
+
+- Append to stream is limited to 7,281 events in a single request ([#77](https://github.com/slashdotdash/eventstore/issues/77)).
+
+### Upgrading
+
+Run the schema migration [v0.13.0.sql](scripts/upgrades/v0.13.0.sql) script against your event store database.
+
 ## v0.12.1
 
 ### Bug fixes